requirements.txt
easydict==1.8
matplotlib==3.0.3
numpy==1.15.4
Pillow==5.4.1
tensorflow-gpu==1.12.0
torch==1.4.0
torchvision==0.5.0
